Move the shunting code to its own specific source file.
