+++ sc2kfix.vcxproj.filters
@@ -470,6 +470,9 @@
     <ClCompile Include="modules\scurk_convert.cpp">
       <Filter>Source Files\Modules</Filter>
     </ClCompile>
+    <ClCompile Include="modules\scurk_objmove.cpp">
+      <Filter>Source Files\Modules</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="sc2kfix.rc">
